refactor(memory): finish legacy package cleanup
