+++ README.md
@@ -150,7 +150,6 @@ in its own module and depends on SDK contracts, never the reverse.
 | [`eval`](eval/)                           | Offline and synthetic quality-evaluation harnesses                             | Workspace module     |
 | [`examples`](examples/)                   | Runnable recall-chatbot and voice-pipeline integrations                        | Examples             |
 | [`tests/conformance`](tests/conformance/) | Provider conformance suites                                                    | Tests                |
-| [`tests/e2e`](tests/e2e/)                 | Retrieval end-to-end coverage                                                  | Tests                |
 
 ---
 
@@ -222,7 +221,6 @@ cd flowcraft
 
 make help          # list every target
 make ci            # vet + test for all in-tree modules
-make test-e2e      # build-tagged retrieval end-to-end suite
 make release-check # validate changesets and the pending release plan
 ```
