+++ examples/.gitignore
@@ -1,3 +1,2 @@
 /voice-pipeline/voice-pipeline
-/chatbot-with-recall/chatbot-with-recall
