+++ memory/knowledge/search.go
@@ -5,11 +5,11 @@ import (
 	"github.com/GizClaw/flowcraft/memory/text/tokenize"
 )
 
-// Knowledge re-exports a minimal slice of [sdk/text] under the
+// Knowledge re-exports a minimal slice of memory/text under the
 // knowledge namespace so existing callers that historically reached
 // for these symbols through the removed sdk/textsearch
 // facade keep working unchanged. The aliases are the documented
-// public surface; new code should import sdk/text/{tokenize,bm25}
+// public surface; new code should import memory/text/{tokenize,bm25}
 // directly.
 
 // Tokenizer aliases [tokenize.Tokenizer].
