+++ sdkx/tool/askuser/doc.go
@@ -10,7 +10,7 @@
 // protocol specs. tool.Tool implementations are concrete adapters
 // — they bridge the generic tool.Tool interface to one specific
 // service — and therefore belong here, mirroring the existing
-// sdk/llm → sdkx/llm/*, sdk/workspace → sdkx/tool/memory, and now
+// sdk/llm → sdkx/llm/*, sdk/workspace → sdkx/tool/anthropicmemory, and now
 // sdk/sandbox → sdkx/tool/exec layouts.
 //
 // # Migration shape
